improve cli usage
