+++ tests/function-controller/testsuite/tests/gitops.go
@@ -1,4 +1,4 @@
-package scenarios
+package tests
 
 import (
 	"fmt"
