+++ tests/function-controller/testsuite/tests/light_runtimes.go
@@ -1,4 +1,4 @@
-package scenarios
+package tests
 
 import (
 	"fmt"
